eps counter values are integer

+++ asab/metrics/metrics.py
@@ -111,7 +111,7 @@ class EPSCounter(Counter):
 
 	def __init__(self, init_values=None):
 		if init_values is not None:
-			init_values = {k: float(v) for k, v in init_values.items()}
+			init_values = {k: int(v) for k, v in init_values.items()}
 
 		super().__init__(init_values=init_values)
 		self.LastTime = time.time()
@@ -127,7 +127,7 @@ class EPSCounter(Counter):
 
 		for field in self.Storage['fieldset']:
 			field['values'] = {
-				k: v / float(delta)
+				k: int(v / float(delta))
 				for k, v in self._actuals.items()
 			}
